Testing file upload system

+++ app/Http/Controllers/Api/UploadController.php
@@ -32,66 +32,76 @@ class UploadController extends Controller
         }
 
         if ($fileType === "video") {
+            $fileName = str_random() . '.' . $file->getClientOriginalExtension();
 
-            $response = [];
-
-
-            $fileName = str_random() . '.'.$file->getClientOriginalExtension();
-
-            $movedFile = $file->move(public_path('/temp/video'), $fileName);
+            $file->move(public_path('/temp/video'), $fileName);
 
             $api = new StreamableApi();
 
+            try {
+                $response = $api->uploadFileFromUrl(
+                    url('/temp/video/' . $fileName)
+                );
+            } catch (\GuzzleHttp\Exception\GuzzleException $e) {
+                unlink(public_path('/temp/video/' . $fileName));
 
-            $response = $api->uploadFileFromUrl(
-                url('/temp/video/' . $fileName)
-            );
+                return "Failed to process video, reason: {$e->getMessage()}";
+            } catch (\Exception $e) {
+                unlink(public_path('/temp/video/' . $fileName));
+
+                return "Failed to process video, reason: {$e->getMessage()}";
+            }
 
             if (isset($response->shortcode)) {
                 $data = $api->getVideoInfo($response->shortcode);
 
+                //Delete the file once streamable has downloaded it
+                if ($data->status === 1) {
+                    unlink(public_path('/temp/video/' . $fileName));
+                }
+
                 return response()->json($data);
 
+            } else {
+                //Remove the file since something must have failed.
+                unlink(public_path('/temp/video/' . $fileName));
             }
+        } else {
+            if ($fileName = Storage::cloud()->putFile('', $file, 'public')) {
 
+                $upload = FileUpload::create([
+                    'user_id'       => $user->id,
+                    'type'          => $file->getClientOriginalExtension(),
+                    'name'          => $randStr,
+                    'file'          => $fileName,
+                    'mime_type'     => $file->getClientMimeType(),
+                    'link'          => Storage::cloud()->url($fileName),
+                    'size_in_bytes' => filesize($file->getPathname()),
+                ]);
 
-            return response()->json($response);
+                if ($fileType === "video") {
 
-            return "done";
+                    $tempThumbnailDir = storage_path() . '/app/public/' . str_random() . '.png';
 
+                    if ($output = shell_exec("ffmpeg -i {$file->getRealPath()} -deinterlace -an -ss 1 -t 00:00:01 -r 1 -y -vcodec mjpeg -f mjpeg {$tempThumbnailDir} 2>&1")) {
+
+                        $thumbnail = Storage::cloud()->putFile('', new \Illuminate\Http\File($tempThumbnailDir),
+                            'public');
+
+                        $upload->thumbnail_url = Storage::cloud()->url($thumbnail);
+                        $upload->save();
+
+                        unlink($tempThumbnailDir);
+                    }
+                }
+
+                return route('file', $upload->name);
+            } else {
+                return "Failed to upload file.";
+            }
         }
 
-        /* if ($fileName = Storage::cloud()->putFile('', $file, 'public')) {
 
-             $upload = FileUpload::create([
-                 'user_id'       => $user->id,
-                 'type'          => $file->getClientOriginalExtension(),
-                 'name'          => $randStr,
-                 'file'          => $fileName,
-                 'mime_type'     => $file->getClientMimeType(),
-                 'link'          => Storage::cloud()->url($fileName),
-                 'size_in_bytes' => filesize($file->getPathname()),
-             ]);
-
-             if ($fileType === "video") {
-
-                 $tempThumbnailDir = storage_path() . '/app/public/' . str_random() . '.png';
-
-                 if ($output = shell_exec("ffmpeg -i {$file->getRealPath()} -deinterlace -an -ss 1 -t 00:00:01 -r 1 -y -vcodec mjpeg -f mjpeg {$tempThumbnailDir} 2>&1")) {
-
-                     $thumbnail = Storage::cloud()->putFile('', new \Illuminate\Http\File($tempThumbnailDir), 'public');
-
-                     $upload->thumbnail_url = Storage::cloud()->url($thumbnail);
-                     $upload->save();
-
-                     unlink($tempThumbnailDir);
-                 }
-             }
-
-             return route('file', $upload->name);
-         } else {
-             return "Failed to upload file.";
-         }*/
     }
 
     /**
